增加调试代码用于调试显示刷新snapshot时会导致Vulkan api卡死的问题。

+++ MediaTimeline.h
@@ -282,7 +282,7 @@ private:
 private:
     bool mQuit;
     float mSnapWidth                {0};
-    uint32_t mSnapHeight            {0};
+    float mSnapHeight               {0};
     float mSnapsInViewWindow        {0};
     int64_t mViewWndStart;
     std::vector<MediaSnapshot::ImageHolder> mSnapImages;
